Information Table Setup and Updated Styles/Layout
Completion of the information table page. This includes a list of the items as well as modals for each item. Data is static right now until the endpoint is complete. I have also put in a dummy search box. I have not figured out how I would like to do the filter yet, so that has been left out. In addition, I updated the banner to display on every page with the page's respective title. Lastly, made some small UI changes to make the site more visually appealing. In addition, I added "fake" images to the home page.